requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic[email]==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0
pytesseract==0.3.10
anthropic==0.18.1
pandas==2.1.4
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
